Add buildIndex method - modify to float datatype

+++ app/src/main/java/lsh/Utils.java
@@ -3,13 +3,13 @@ package lsh;
 public class Utils {
     
 
-    public static double dot(double[] aVec, double[] bVec){
+    public static float dot(float[] aVec, float[] bVec){
 
         if (aVec.length != bVec.length) {
             throw new IllegalArgumentException("Vectors of unequal length passed");
         }
 
-        double dotProduct = 0.0;
+        float dotProduct = 0.0f;
         
         for (int i = 0; i < aVec.length; i++) {
             dotProduct += aVec[i] * bVec[i];
@@ -18,11 +18,11 @@ public class Utils {
         return dotProduct;
     }
 
-    public static double euclideanDistance(double[] aVec, double[] bVec) {
+    public static float euclideanDistance(float[] aVec, float[] bVec) {
 
         // To be implemented
 
-        return -1.0;
+        return -1.0f;
     }
 
     //NEW METHOD HERE
